feat(schema): allow wasAttributedTo to be a SoftwareApplication
The prov schema says that the value of wasAttributedTo can be any
agent, but it can currently only be a LegalPersonInterface.
We want to use this property to indicate that a node was created
by a particular software application. To do this, introduce an
AgentInterface which is implemented by SoftwareApplication and the
types that already implemented LegalPersonInterface.