feat(schema): Allow custom motivations to be created
A user should be able to make a motivation that more accurately
describes one of the existing oa motivations. This AnnotationCEMotivation
should have `broaderMotivation` set to an existing motivation enum
value, or have `broaderUrl` set to an external URL describing another
motivation.
DefinedTerms which are also motivations can also be set to one of
the existing motivation enum values.

Annotations can have a motivation set to an enum value, one of these
custom nodes, or an external URL.